refactor: consolidate lib, styles, and fonts into app directory
Move all shared resources into app directory for better organization:
- src/lib → app/lib (utilities and data fetching)
- styles → app/styles (CSS modules and global styles)
- src/fonts → app/fonts (font files)

Update all import paths across the application accordingly.
Remove empty src directory.
