+++ app/blog/[slug]/page.tsx
@@ -1,8 +1,8 @@
 import { FetchBlocks } from "rotion";
 import { notFound } from "next/navigation";
 import PageDetail from "../../../components/page-detail";
-import { GetBlog, GetPaths } from "../../../src/lib/blog";
-import { MakeOgImage } from "../../../src/lib/ogimage";
+import { GetBlog, GetPaths } from "../../lib/blog";
+import { MakeOgImage } from "../../lib/ogimage";
 
 type Params = {
 	slug: string;
