+++ app/blog/components/blog-list.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { UsePagination } from "../components/rotion-wrappers";
 import Hed from "../components/hed";
-import type { Blog } from "../../../src/lib/blog";
+import type { Blog } from "../../lib/blog";
 
 type Props = {
 	pages: Blog[];
