+++ app/fonts/NotoSansJP-Black.woff2
