+++ app/layout.tsx
@@ -1,12 +1,12 @@
 import "rotion/style.css";
-import "../styles/globals.css";
+import "./styles/globals.css";
 import "normalize.css/normalize.css";
 import type { Metadata } from "next";
 import { Suspense } from "react";
 import Footer from "./components/footer";
 import GA from "./components/ga";
 import Header from "./components/header";
-import { notosans, notoserif } from "../src/lib/fonts";
+import { notosans, notoserif } from "./lib/fonts";
 
 export const metadata: Metadata = {
 	alternates: {
