+++ app/projects/page.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
 import Link from "next/link";
 import Hed from "../components/hed";
-import { MakeOgImage } from "../../src/lib/ogimage";
-import { GetProjects } from "../../src/lib/project";
-import Styles from "../../styles/Project.module.css";
+import { MakeOgImage } from "../lib/ogimage";
+import { GetProjects } from "../lib/project";
+import Styles from "../styles/Project.module.css";
 
 const title = "Project";
 const desc =
